requirements.txt
django==4.1.2
eventlet==0.33.1
python-engineio==4.0.0
python-socketio==5.0.4
pytz==2018.7
six==1.10.0
docker==6.0.1